Image_to_Recipe_Generator/requirements.txt
# List of required packages for the applica
# Core ML and AI Libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
google-generativeai==0.3.1

# Web Frameworks
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.2
python-multipart==0.0.6

# Utility Libraries
python-dotenv==1.0.0
pillow==9.5.0
requests==2.31.0
langchain==0.0.335

# Data Processing and NLP
spacy==3.7.2
nltk==3.8.1
numpy==1.24.3

# Optional visualization and analysis
matplotlib==3.7.2
seaborn==0.12.2

# Development and Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0